+++ tracer/src/Datadog.Trace.ClrProfiler.Managed.Loader/Startup.cs
@@ -78,16 +78,19 @@ namespace Datadog.Trace.ClrProfiler.Managed.Loader
 
                 StartupLogger.Debug("Resolved Datadog.Trace.dll TFM directory to: {0}", ManagedProfilerDirectory);
 
-                // Publish the resolver's state BEFORE subscribing the handler. Reading/writing a
-                // static field of ManagedProfilerAssemblyResolver triggers its (trivial) type init
-                // here on the main thread, so by the time any ThreadPool thread later dispatches
-                // the handler, the type is already fully initialized and the dispatch doesn't have
-                // to wait on Startup..cctor. See APMS-19239.
-                ManagedProfilerAssemblyResolver.ManagedProfilerDirectory = ManagedProfilerDirectory;
-
                 try
                 {
+#if NETFRAMEWORK
+                    // On .NET Framework, route AssemblyResolve through a class other than Startup so
+                    // the handler doesn't require Startup..cctor to have finished. If a configBuilder
+                    // on <appSettings> issues sync-over-async work during Startup..cctor, the async
+                    // continuation may fire AssemblyResolve on a ThreadPool thread; a handler on
+                    // Startup itself would deadlock waiting on Startup..cctor. See APMS-19239.
+                    ManagedProfilerAssemblyResolver.ManagedProfilerDirectory = ManagedProfilerDirectory;
                     AppDomain.CurrentDomain.AssemblyResolve += ManagedProfilerAssemblyResolver.OnAssemblyResolve;
+#else
+                    AppDomain.CurrentDomain.AssemblyResolve += AssemblyResolve_ManagedProfilerDependencies;
+#endif
                 }
                 catch (Exception ex)
                 {
@@ -97,7 +100,7 @@ namespace Datadog.Trace.ClrProfiler.Managed.Loader
 #if NETCOREAPP
                 try
                 {
-                    System.Runtime.Loader.AssemblyLoadContext.Default.Resolving += ManagedProfilerAssemblyResolver.OnAssemblyLoadContextResolving;
+                    System.Runtime.Loader.AssemblyLoadContext.Default.Resolving += (_, assemblyName) => ResolveAssembly(assemblyName.Name);
                 }
                 catch (Exception ex)
                 {
@@ -190,7 +193,11 @@ namespace Datadog.Trace.ClrProfiler.Managed.Loader
                 // We will try to resolve it manually as a last chance.
                 StartupLogger.Log(ex, "Error on assembly load: {0}, Trying to solve it manually...", assemblyString);
 
+#if NETFRAMEWORK
                 var assembly = ManagedProfilerAssemblyResolver.ResolveAssembly(assemblyString);
+#else
+                var assembly = ResolveAssembly(assemblyString);
+#endif
                 if (assembly is not null)
                 {
                     StartupLogger.Log("Assembly '{0}' was resolved manually.", assemblyString);
